refactoring: move reusable functions out of run.py
